+++ src/Symfony/Component/HttpFoundation/CHANGELOG.md
@@ -22,6 +22,7 @@ CHANGELOG
 ---
 
  * Add the `litespeed_finish_request` method to work with Litespeed
+ * Deprecate `upload_progress.*` and `url_rewriter.tags` session options
 
 5.3
 ---